+++ CHANGELOG.md
@@ -25,6 +25,9 @@ In addition to these necessary markup changes, the bootstrap_form API itself has
 ### Bugfixes
 
 * Your contribution here!
+* [#357](https://github.com/bootstrap-ruby/bootstrap_form/pull/357) if provided,
+  use html option `id` to specify `for` attribute on label
+  [@duleorlovic](https://github.com/duleorlovic)
 
 ## [2.7.0][] (2017-04-21)